requirements.txt
Django==3.2.4
django-environ==0.4.5
djangorestframework==3.12.4
mysqlclient==2.0.3
cryptography~=2.8
django-cors-headers~=3.1.1
drf-jwt~=1.13.3
pyjwt~=1.7.1
requests~=2.22.0

